Made each note have a posMode, control on change only now possible.
To avoid conflicting controls from knobs and MIDI messages, some functions take
in the last value that was supposed to be set. If it is the same as what was
attempted to be set before, the function returns. This is so that the switch can
be in one setting and then a MIDI message can be received that can change it.
Even though it will be seen that the switch is in one position, it will not
change the setting because it will not call the function until what is requested
is different from what was requested last time.

Fine pitch control and stride state control have been added, too. Need to change
many synth control functions to accept the parameter set they are manipulating.