Even if pitch knob moved while UNI pressed on note 1 (N1/BAR), the pitch is controlled as normal (all pitches changed, but normally only one pitch is available for playing, others must be accessed using MIDI).
In FREE recording mode, playback is not retriggered after recording.

In AREC mode, if only playing, when the record switch is pressed, recordings
begin again with the next beat. The FBK switch does nothing in this case.
Recordings are requested as in the other modes (but with tempo adjustment and
re-scheduling based on the record time) when not playing.

None of these have yet been tested.