Updated board tests
Updated tests for switches, pots and LEDs. They must eventually be updated to
reflect the new layout of the switches.

When building using make or running the tests, the defines BOARD_VERSION and
OPENOCD_INTERFACE have been made available.
BOARD_VERSION allows you to specify the version of the circuit so that ports are
addressed properly and the codec is setup correctly.
OPENOCD_INTERFACE lets you choose what version of the development board you are
using to flash the device.