Pitches can now be updated every signal processing tick
Preliminary tests show it works. The following has been tested:

The changing of pitches using the knobs.
The resetting of pitches using the UNI switch and when recording loops.

What needs to be tested:

The reading of pitches as "absolute" when triggered from a MIDI note on event.
In this case they should ignore the rate bus and use a fixed value.