Pointer to a gain value now updates the gains
This is so they can be adjusted in real time.

One problem is that if the amplitude is too low, no note is triggered, which is okay
except that if you think you are fading in a note with the knob, it won't be sounding
... probably okay.

Need to do updates for MIDI, not yet done.