Footswitches now cause instant action
When the footswitch is pressed, it waits some number of checks while down.
If it has remained down for this many checks, the action happens. This makes
it so when the footswitch is pressed down, the action happens, not when it comes
up (like before). The other momentary switches still behave as before
(where the action happens when the switch is released).

To do this change, a state field was added to the switch debouncer and the
debounce style is chosen upon its initialization.

So far the switches seem to behave well.