Working on cleaned up MIDI
Have gone through until the end of the N2 controls

Described correctly the stride reset for N2 and N3.

You can control intermittency with an exact value now through MIDI using
synth_control_set_intermittency_idx.

Updated scripts/test_midi to use new patterns (basically we're getting rid of
the convoluted MIDI map and just listing all the functions individually).

set_offset was wrong if you changed the eventDeltaBeats of N1 through MIDI. Now
it is just a value between 0 and 1 as it should be. It could be a value between
0 and intermittency+1, maybe later...

If for some reason the stored midi channel was invalid, correct it to
be 0 and store that midi channel.

Set number of repeats exactly using MIDI.