sustain time should now correctly be computed in seconds
the error was that the value of the attack and the decay times was used as a
fraction of the total length, which is wrong. At this point in the code they
represented seconds.

But we still get a click for no fade / attack time. Perhaps this is due to rounding...