tempo scaling only applied when switch in QUANT position
when N1 is selected and the recording mode is FREE, setting the
FREE/QUANT/SKIP switch to QUANT allows selecting tempo scaling. Now,
when QUANT is not the current setting, no scaling is applied. When the
switch returns to the QUANT setting, the scaling is applied again.

TODO: Need to test how this interacts with the MIDI setting of tempo
scaling.